agents/airbnb_planner_multiagent-main/pyproject.toml
# Copyright 2026 Tsinghua University
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# This file was created by Tsinghua University and is not part of
# the original agentgateway project by Solo.io.

[project]
name = "airbnb_planner_multiagent"
version = "0.1.0"
description = "Plan a trip using a multi-agent system communicating over A2A and using MCP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.0",
    "geopy>=2.4.1",
    "google-adk>=1.7.0",
    "gradio>=5.30.0",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-google-vertexai>=2.0.24",
    "langgraph>=0.4.5",
    "mcp>=1.5.0",
    "a2a-sdk>=0.3.0",
    "litellm",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "ruff>=0.12.8",
]
